requirements.txt
# LangChain Core
langchain~=0.3.26
langchain-core

# OpenAI Integration
langchain-openai~=0.3.23
openai~=1.86.0

# Anthropic Integration
langchain-anthropic

# Google Gemini (PaLM) Integration
langchain-google-genai
google-generativeai

# Hugging Face Integration
langchain-huggingface
transformers
huggingface-hub

# Environment Variable Management
python-dotenv~=1.1.0

# Machine Learning Utilities
numpy~=2.3.0
scikit-learn~=1.7.0
sentence-transformers~=4.1.0
python-docx~=1.1.2
langchain-community~=0.3.26